scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.2",
    "web3>=7.14.1",
    "apscheduler>=3.10.0",
]
